+++ mindbox_android/example/lib/main.dart
@@ -31,7 +31,7 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future<void> initPlatformState() async {
-    String sdkVersion = await MindboxPlatform.instance.sdkVersion;
+    String sdkVersion = await MindboxPlatform.instance.nativeSdkVersion;
 
     if (!mounted) return;