+++ mindbox_ios/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0
+
+* Upgrade native SDK dependency to v2.0.1.
+* Add support of Huawei Push Kit.
+* Update onPushClickReceived method: add second argument with push payload.
+
 ## 1.0.2
 
 * Add iOS old format push notifications support.